requirements.txt
black==22.12.0
bleach==6.0.0
build==0.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.3
decorator==4.4.2
docutils==0.19
flake8==4.0.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.2.3
keyring==23.13.1
markdown-it-py==2.2.0
mccabe==0.6.1
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mypy==0.942
mypy-extensions==1.0.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pathspec==0.11.1
pep517==0.13.0
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.0
pluggy==1.0.0
proglog==0.1.10
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.0.0
readme-renderer==37.3
requests==2.29.0
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.5
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
twine==4.0.2
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
zipp==3.15.0
